requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-countries==7.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.5.0
pluggy==1.0.0
pycparser==2.21
pyOpenSSL==23.2.0
pytest==7.3.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
